Add support for the case of split assemblies
In the case of v4 - and maybe others - one assembly was split up and types were moved into separate assemblies. The previous way of migrating assemblies was to just rename it, but if the types were in separate assemblies, the same assembly would just be renamed multiple times. This way adds new assemblies and removes any old ones.